+++ docs/changelog.md
@@ -5,3 +5,4 @@
 ### New
 
 * `build-contracts` - convert PHPUnit testdox.txt into Markdown
+* `build-php-badges` - get PHP badges from shields.io